--- data/apis/amberdata/deployments/220623-2000/airkeeper/aws/secrets.json
@@ -1,1 +1,0 @@
-{ "filename": "secrets.env", "content": "" }
