+++ package.json
@@ -16,30 +16,30 @@
     "registry": "https://registry.npmjs.org/"
   },
   "scripts": {
-    "create-boilerplate": "ts-node src/create-boilerplate.ts",
-    "create-config": "ts-node src/create-config.ts",
-    "create-airseeker-config": "ts-node src/create-airseeker-config.ts",
-    "export-ops-json-payloads": "ts-node src/export-ops-json-payloads.ts",
-    "import-receipt": "ts-node src/import-receipt.ts",
+    "create-boilerplate": "ts-node src/deployment/create-boilerplate.ts",
+    "create-config": "ts-node src/deployment/create-config.ts",
+    "create-airseeker-config": "ts-node src/deployment/create-airseeker-config.ts",
+    "export-ops-json-payloads": "ts-node src/utils/export-ops-json-payloads.ts",
+    "import-receipt": "ts-node src/utils/import-receipt.ts",
     "eth:deploy": "cd chain && hardhat deploy --network $NETWORK",
     "deploy-all": "yarn deploy-airnode; yarn deploy-airkeeper",
-    "deploy-airkeeper": "yarn ts-node src/deploy-airkeeper",
-    "deploy-airnode": "yarn ts-node src/deploy-airnode",
-    "remove-airnode": "yarn ts-node src/remove-airnode",
-    "generate-gateway-key": "ts-node src/generate-gateway-key.ts",
-    "generate-topup-wallets": "ts-node src/generate-topup-wallets.ts",
-    "register-subscriptions": "ts-node src/register-subscriptions.ts",
-    "add-missing-beacon-metadata": "ts-node src/add-missing-beacon-metadata.ts",
+    "deploy-airkeeper": "yarn ts-node src/deployment/deploy-airkeeper",
+    "deploy-airnode": "yarn ts-node src/deployment/deploy-airnode",
+    "remove-airnode": "yarn ts-node src/deployment/remove-airnode",
+    "generate-gateway-key": "ts-node src/utils/generate-gateway-key.ts",
+    "generate-topup-wallets": "ts-node src/utils/generate-topup-wallets.ts",
+    "register-subscriptions": "ts-node src/register/register-subscriptions.ts",
+    "add-missing-beacon-metadata": "ts-node src/explorer/add-missing-beacon-metadata.ts",
     "build": "yarn run clean && yarn run compile",
-    "check-subscriptions": "ts-node src/check-subscriptions.ts",
-    "check-dapi-name": "ts-node src/check-dapi-name.ts",
-    "check-policy-readers": "ts-node src/check-policy-readers.ts",
-    "enable-policy-readers": "ts-node src/enable-policy-readers.ts",
+    "check-subscriptions": "ts-node src/inspect/check-subscriptions.ts",
+    "check-dapi-name": "ts-node src/inspect/check-dapi-name.ts",
+    "check-policy-readers": "ts-node src/inspect/check-policy-readers.ts",
+    "enable-policy-readers": "ts-node src/register/enable-policy-readers.ts",
     "clean": "rimraf -rf ./dist *.tgz",
     "compile": "tsc -p tsconfig.json",
-    "normalize-and-validate": "ts-node src/normalize-and-validate.ts",
-    "validate": "ts-node src/validate.ts",
-    "normalize": "ts-node src/normalize.ts",
+    "normalize-and-validate": "ts-node src/validation/normalize-and-validate.ts",
+    "validate": "ts-node src/validation/validate.ts",
+    "normalize": "ts-node src/validation/normalize.ts",
     "lint": "yarn run prettier:check && yarn run lint:eslint",
     "lint:eslint": "eslint . --ext .js,.ts",
     "prettier:check": "prettier --check \"./**/*.{js,ts,md,json}\"",
