+++ src/deployment/create-boilerplate.ts
@@ -1,11 +1,11 @@
 import { join } from 'path';
 import { PromptObject } from 'prompts';
-import { runAndHandleErrors } from './utils/cli';
-import { sanitiseFilename } from './utils/filesystem';
-import { emptyObject } from './utils/normalization';
-import { promptQuestions } from './utils/prompts';
-import { readOperationsRepository } from './utils/read-operations';
-import { writeOperationsRepository } from './utils/write-operations';
+import { runAndHandleErrors } from '../utils/cli';
+import { sanitiseFilename } from '../utils/filesystem';
+import { emptyObject } from '../validation/normalization';
+import { promptQuestions } from '../utils/prompts';
+import { readOperationsRepository } from '../utils/read-operations';
+import { writeOperationsRepository } from '../utils/write-operations';
 
 const questions: PromptObject[] = [
   {
