+++ src/register/enable-policy-readers.ts
@@ -1,10 +1,10 @@
 import { NonceManager } from '@ethersproject/experimental';
 import { ContractTransaction, ethers } from 'ethers';
-import { Policy } from './types';
-import { runAndHandleErrors } from './utils/cli';
-import { getDapiNameHash, getDapiServerContract } from './utils/evm';
-import { loadCredentials } from './utils/filesystem';
-import { readAndValidateOperationsRepository } from './utils/read-operations';
+import { Policy } from '../validation/types';
+import { runAndHandleErrors } from '../utils/cli';
+import { getDapiNameHash, getDapiServerContract } from '../utils/evm';
+import { loadCredentials } from '../utils/filesystem';
+import { readAndValidateOperationsRepository } from '../utils/read-operations';
 
 const printTransactionEvents = (tx: ethers.ContractReceipt, chainId: number) => {
   tx.events
