+++ src/utils/generate-gateway-key.ts
@@ -1,9 +1,9 @@
 import { randomBytes } from 'crypto';
-import { runAndHandleErrors } from './utils/cli';
+import { runAndHandleErrors } from './cli';
 
 const main = async () => {
-  const httpGatewayKey = randomBytes(48).toString('hex');
-  const signedHttpGatewayKey = randomBytes(48).toString('hex');
+  const httpGatewayKey = randomBytes(32).toString('hex');
+  const signedHttpGatewayKey = randomBytes(32).toString('hex');
 
   console.log(`🔑 Generated HTTP Gateway API Key: ${httpGatewayKey}`);
   console.log(`🔑 Generated HTTP Signed Data Gateway API Key: ${signedHttpGatewayKey}`);
