+++ src/utils/index.ts
@@ -2,5 +2,3 @@ export * as cli from './cli';
 export * as writeOperations from './write-operations';
 export * as readOperations from './read-operations';
 export * as filesystem from './filesystem';
-export * as normalization from './normalization';
-export * as validation from './validation';
