+++ src/validation/airseeker-validation.ts
