+++ src/validation/normalization.ts
@@ -3,9 +3,9 @@ import { join } from 'path';
 import { ethers } from 'ethers';
 import { encode } from '@api3/airnode-abi';
 import { parse } from 'dotenv';
-import { sanitiseFilename } from './filesystem';
-import { readJsonFile } from './read-operations';
-import { OperationsRepository, Secrets, ChainDeploymentReferences } from '../types';
+import { sanitiseFilename } from '../utils/filesystem';
+import { readJsonFile } from '../utils/read-operations';
+import { OperationsRepository, Secrets, ChainDeploymentReferences } from './types';
 
 export const normalize = (payload: OperationsRepository) => {
   const apis = Object.fromEntries(
