+++ src/validation/validation.ts
@@ -13,7 +13,7 @@ import {
   validatePoliciesDatafeedReferences,
   validateTemplatesEndpointIdReferences,
 } from './validation-refinements';
-import { OperationsRepository, ValidationResult } from '../types';
+import { OperationsRepository, ValidationResult } from './types';
 
 const { oisSchema } = ois;
