+++ test/generate-topup-wallets.test.ts
@@ -6,8 +6,8 @@ import { deriveWalletPathFromSponsorAddress } from '@api3/airnode-node/dist/src/
 import { readOperationsRepository } from '../src/utils/read-operations';
 import { PROTOCOL_ID_PSP } from '../src/utils/evm';
 import { writeOperationsRepository } from '../src/utils/write-operations';
-import { generateTopupWallets } from '../src/generate-topup-wallets';
-import { OperationsRepository } from '../src/types';
+import { generateTopupWallets } from '../src/utils/generate-topup-wallets';
+import { OperationsRepository } from '../src/validation/types';
 
 describe('generate-topup-wallets', () => {
   const tempTestPath = join(__dirname, '../temporary_test_folder');
