+++ test/utils/normalize.test.ts
@@ -1,9 +1,9 @@
 import { join } from 'path';
 import { ethers } from 'ethers';
 import { encode } from '@api3/airnode-abi';
-import { OperationsRepository } from '../../src/types';
+import { OperationsRepository } from '../../src/validation/types';
 import { readOperationsRepository } from '../../src/utils/read-operations';
-import { normalize } from '../../src/utils/normalization';
+import { normalize } from '../../src/validation/normalization';
 
 describe('normalize', () => {
   let unsanitizedMockData: OperationsRepository;
